+++ modules/core/plan/src/Entity/PlanRecordTypeInterface.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Drupal\plan\Entity;
+
+use Drupal\Core\Config\Entity\ConfigEntityInterface;
+use Drupal\Core\Entity\EntityDescriptionInterface;
+
+/**
+ * Provides an interface for defining Plan record relationship type entities.
+ */
+interface PlanRecordTypeInterface extends ConfigEntityInterface, EntityDescriptionInterface {
+
+}